Fixing more warnings (#5189)
* classes need to be in new files

* adding in new lines
