+++ src/core/Microsoft.PowerToys.Settings.UI/Controls/GroupTitleTextBlock.xaml.cs
@@ -19,7 +19,6 @@ using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Navigation;
 
 // The User Control item template is documented at https://go.microsoft.com/fwlink/?LinkId=234236
-
 namespace Microsoft.PowerToys.Settings.UI.Controls
 {
     public sealed partial class GroupTitleTextBlock : UserControl
